// types/types.ts
export type vehicleData = {
  id: string
  name: string
  logo: string
  price: string
  carType: string
  carPass: string
  personLogo: string
  topSpeed: string
  carBrand: string
  show_room: string
  transmission: string
  description: string
  ownerName: string
  ownerNumber: number
  doorCount: number
  HorsePower: number
}